+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @varveai/adit-hooks
 
+## 0.3.6
+
+### Patch Changes
+
+- 部分代码重构
+- Updated dependencies
+  - @varveai/adit-engine@0.3.6
+  - @varveai/adit-cloud@0.3.6
+  - @varveai/adit-core@0.3.6
+
+## 0.3.5
+
+### Patch Changes
+
+- 部分代码重构
+- Updated dependencies
+  - @varveai/adit-engine@0.3.5
+  - @varveai/adit-cloud@0.3.5
+  - @varveai/adit-core@0.3.5
+
+## 0.3.4
+
+### Patch Changes
+
+- 获取 git 记录方式调整
+- Updated dependencies
+  - @varveai/adit-engine@0.3.4
+  - @varveai/adit-cloud@0.3.4
+  - @varveai/adit-core@0.3.4
+
+## 0.3.3
+
+### Patch Changes
+
+- codex 支持
+- Updated dependencies
+  - @varveai/adit-engine@0.3.3
+  - @varveai/adit-cloud@0.3.3
+  - @varveai/adit-core@0.3.3
+
 ## 0.3.2
 
 ### Patch Changes